bip39: consolidate/remove unnecessary functions/files
-use 'in-house' utility functions to reduce imports
-remove unneeded headers/circular dependencies
