+++ Makefile.am
@@ -13,17 +13,7 @@ $(LIBSECP256K1): $(wildcard src/secp256k1/src/*) $(wildcard src/secp256k1/includ
 lib_LTLIBRARIES = libdogecoin.la
 include_HEADERS = include/dogecoin/libdogecoin.h
 noinst_HEADERS = \
-    include/bip39/chinese_simplified.h \
-    include/bip39/chinese_traditional.h \
-    include/bip39/czech.h \
-    include/bip39/english.h \
-    include/bip39/french.h \
     include/bip39/index.h \
-    include/bip39/italian.h \
-    include/bip39/japanese.h \
-    include/bip39/korean.h \
-    include/bip39/portuguese.h \
-    include/bip39/spanish.h \
     include/dogecoin/address.h \
     include/dogecoin/aes.h \
     include/dogecoin/base58.h \
@@ -64,10 +54,7 @@ libdogecoin_la_SOURCES = \
     src/aes.c \
     src/base58.c \
     src/bip32.c \
-    src/bip39/bip39c.c \
-    src/bip39/bip39.c \
-    src/bip39/crypto.c \
-    src/bip39/conversion.c \
+    src/bip39.c \
     src/bip44.c \
     src/block.c \
     src/buffer.c \
