--- src/bip39/conversion.h
@@ -1,44 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2018 David L. Whitehurst
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- * conversion.h (header)
- * Reusable conversion (beyond casting) functions.
- *
- * author: David L. Whitehurst
- * date: June 8, 2018
- *
- * Find this code useful? Please donate:
- *  Bitcoin: 1Mxt427mTF3XGf8BiJ8HjkhbiSVvJbkDFY
- *
- */
-
-#ifndef BIP39C_CONVERSION_H
-#define BIP39C_CONVERSION_H
-
-
-/*
- * Function declarations
- */
-unsigned char *hexstr_to_char(const char *hexstr);
-
-#endif //BIP39C_CONVERSION_H
