--- src/bip39/crypto.c
@@ -1,64 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2018 David L. Whitehurst
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- * crypto.c (source)
- * Current cryptographic utility functions using OpenSSL low-level code.
- *
- * author: David L. Whitehurst
- * date: June 8, 2018
- *
- * Find this code useful? Please donate:
- *  Bitcoin: 1Mxt427mTF3XGf8BiJ8HjkhbiSVvJbkDFY
- *
- */
-
-#include "crypto.h"
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include <dogecoin/sha2.h>
-
-/*
- * This function implements a SHA256 checksum and fills a
- * 64-byte buffer with a fixed-length hex representation of the
- * message digest.
- */
-
-int sha256(char *string, char outputBuffer[65]) {
-    sha256_context ctx;
-    uint8_t hash[SHA256_DIGEST_LENGTH];
-
-    sha256_init(&ctx);
-    sha256_write(&ctx, (const uint8_t*) string, strlen(string));
-    sha256_finalize(&ctx, hash);
-
-    for (size_t i = 0; i < SHA256_DIGEST_LENGTH; i++) {
-        sprintf(outputBuffer + (i * 2), "%02hhX ", hash[i]);
-    }
-
-    outputBuffer[64] = 0;
-    return 0;
-}
-
